api_service/requirements.txt
fastapi
uvicorn
requests
python-dotenv
celery
redis